Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/2dcontext/fill-and-stroke-styles/2d.fillStyle.parse.current.changed-expected.txt

2d.fillStyle.parse.current.changed
currentColor is computed when the attribute is set, not when it is painted
Actual output:
Expected output:


PASS currentColor is computed when the attribute is set, not when it is painted 
